Added: -Админ панель  -CRUD of categories  -RUD of orders  -images(UPLOAD and Getting)  Думаю на этом все, в будущем возможно буду доделывать,косяков много,но пока что на этом все.  Был рад иметь мой первый полный и осознанный опыт с MVC паттерном,надеюсь в будущем я стану намного лучше в плане кода,документации и архитектуры, и конечно же технологий.

+++ Php-Start_Project/controllers/AdminProdController.php
@@ -7,9 +7,17 @@ class AdminProdController {
             return true;
     }
 
+
     public function actionAddProd(){
         $categories = Category::getCategoryList();
         $options = [];
+        $altImg = "view/upload/images/products/АльтРубашка.jpg";
+
+        $Prods = Product::getProducts();
+        $id = 0;
+        foreach ($Prods as $prod){
+            $id = $prod['id'];
+        }
         if (isset($_POST['submit'])) {
 
             $options['article'] = $_POST['article'];
@@ -24,7 +32,17 @@ class AdminProdController {
             $options['rec'] = $_POST['rec'];
             $options['hid'] = $_POST['hid'];
 
+            $id++;
+            if (is_uploaded_file($_FILES['image']['tmp_name'])) {
+                move_uploaded_file($_FILES['image']['tmp_name'],
+               ROOT."/view/upload/images/products/{$id}.jpg");
+                $path ="view/upload/images/products/{$id}.jpg";
+            }
+
+
             Product::addProduct($options);
+
+            header("Location:/PHP-start_PracticeE-martCreating/Php-Start_Project/index.php/aProd");
 
 //                foreach ($options as $option){
 //                    echo $option."<br>";
@@ -39,6 +57,11 @@ class AdminProdController {
 
         $categories = Category::getCategoryList();
         $product = Product::getProductById($id);
+
+        $path ="view/upload/images/products/{$id}.jpg";
+        $altImg = "view/upload/images/products/АльтРубашка.jpg";
+        echo $path;
+
         $options = [];
             if (isset($_POST['submit'])) {
 
@@ -48,13 +71,17 @@ class AdminProdController {
                 $options['category'] = $_POST['category'];
                 $options['price'] = $_POST['price'];
                 $options['brand'] = $_POST['brand'];
-                //$options['image'] = $_POST['image'];
                 $options['description'] = $_POST['description'];
                 $options['new'] = $_POST['new'];
                 $options['rec'] = $_POST['rec'];
                 $options['hid'] = $_POST['hid'];
 
+                if (is_uploaded_file($_FILES['image']['tmp_name'])) {
+                    move_uploaded_file($_FILES['image']['tmp_name'],
+                        ROOT."/view/upload/images/products/{$id}.jpg");
+                }
         Product::updateProduct($id,$options);
+                header("Location:/PHP-start_PracticeE-martCreating/Php-Start_Project/index.php/aProd");
 
 //                foreach ($options as $option){
 //                    echo $option."<br>";
@@ -66,9 +93,15 @@ class AdminProdController {
         return true;
     }
 
-    public function actionDelete($id){
-
-        Product::deleteProduct($id);
-        header("Location:".$_SERVER['HTTP_REFERER']);
+    public function actionDelete($id)
+    {
+        if (isset($_POST['yes'])) {
+            Product::deleteProduct($id);
+            header("Location:/PHP-start_PracticeE-martCreating/Php-Start_Project/index.php/aProd");
+        } elseif(isset($_POST['no'])) {
+            header("Location:/PHP-start_PracticeE-martCreating/Php-Start_Project/index.php/aProd");
+        }
+        require_once 'view/admin/prodDel.php';
+        return true;
     }
 }
